+++ src/wadogo/scale/interpolated.clj
@@ -1,26 +1,32 @@
 (ns wadogo.scale.interpolated
-  (:require [fastmath.interpolation :refer [interpolators-1d-list]]
+  (:require [fastmath.interpolation :as i]
             
             [wadogo.common :refer [scale ->ScaleType strip-keys merge-params]]))
-
-(def ^:private interpolators
-  {:cubic :cubic-spline
-   :kriging :kriging-spline
-   :linear :linear-smile})
 
 (defmethod scale :interpolated
   ([_ ] (scale :interpolated {}))
   ([s params]
    (let [params (merge-params s params)
          pairs (map vector (:domain params) (:range params))
-         interpolator-key (get interpolators (:interpolator params) (:interpolator params))
-         interpolator (apply partial (interpolators-1d-list interpolator-key) (:interpolator-params params))
          fpairs (sort-by first pairs)
-         forward (interpolator (map first fpairs)
-                               (map second fpairs))
+         interpolator (:interpolator params :linear)
+         interpolator-params (:interpolator-params params)
+         forward (if (fn? interpolator)
+                   (apply interpolator
+                          (map first fpairs) (map second fpairs)
+                          interpolator-params)
+                   (apply i/interpolation interpolator
+                          (map first fpairs) (map second fpairs)
+                          interpolator-params))
          ipairs (sort-by second pairs)
-         inverse (interpolator (map second ipairs)
-                               (map first ipairs))]
+         inverse (if (fn? interpolator)
+                   (apply interpolator
+                          (map first fpairs) (map second fpairs)
+                          interpolator-params)
+                   (apply i/interpolation interpolator
+                          (map second ipairs)
+                          (map first ipairs)
+                          interpolator-params))]
      (->ScaleType :interpolated (:domain params) (:range params) (:ticks params) (:formatter params)
                   forward
                   inverse